+++ Fabric/src/main/java/com/cerbon/better_totem_of_undying/fabric/charm/CharmIntegration.java
@@ -1,5 +1,7 @@
 package com.cerbon.better_totem_of_undying.fabric.charm;
 
+import com.cerbon.better_totem_of_undying.BetterTotemOfUndying;
+import com.cerbon.better_totem_of_undying.config.BTUConfigs;
 import com.mojang.math.Axis;
 import dev.emi.trinkets.api.client.TrinketRenderer;
 import dev.emi.trinkets.api.client.TrinketRendererRegistry;
@@ -15,10 +17,13 @@ public class CharmIntegration {
     public static void renderVoidTotemOnChest() {
         TrinketRendererRegistry.registerRenderer(Items.TOTEM_OF_UNDYING, (stack, slotReference, contextModel, matrices, vertexConsumers, light, entity, limbAngle, limbDistance, tickDelta, animationProgress, headYaw, headPitch) -> {
 
-            if (entity instanceof AbstractClientPlayer player){
+            if (entity instanceof AbstractClientPlayer player) {
                 TrinketRenderer.translateToChest(matrices, (PlayerModel<AbstractClientPlayer>) contextModel, player);
 
+                BTUConfigs btuConfigs = BetterTotemOfUndying.config;
+
                 matrices.scale(0.35F, 0.35F, 0.35F);
+                matrices.translate(btuConfigs.charm.xOffset, btuConfigs.charm.yOffset, btuConfigs.charm.zOffset);
                 matrices.mulPose(Axis.ZP.rotationDegrees(180));
 
                 Minecraft.getInstance().getItemRenderer().renderStatic(stack, ItemDisplayContext.FIXED, 15728880, OverlayTexture.NO_OVERLAY, matrices, vertexConsumers, Minecraft.getInstance().level, 0);