+++ Forge/src/main/java/com/cerbon/better_totem_of_undying/forge/client/RenderTotemOnChest.java
@@ -1,5 +1,7 @@
 package com.cerbon.better_totem_of_undying.forge.client;
 
+import com.cerbon.better_totem_of_undying.BetterTotemOfUndying;
+import com.cerbon.better_totem_of_undying.config.BTUConfigs;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.EntityModel;
@@ -14,7 +16,7 @@ import org.jetbrains.annotations.NotNull;
 import top.theillusivec4.curios.api.SlotContext;
 import top.theillusivec4.curios.api.client.ICurioRenderer;
 
-public class RenderVoidTotemOnChest implements ICurioRenderer {
+public class RenderTotemOnChest implements ICurioRenderer {
 
     @Override
     public <T extends LivingEntity, M extends EntityModel<T>> void render(ItemStack itemStack, @NotNull SlotContext slotContext, PoseStack poseStack, RenderLayerParent<T, M> renderLayerParent, MultiBufferSource multiBufferSource, int i, float v, float v1, float v2, float v3, float v4, float v5) {
@@ -22,8 +24,10 @@ public class RenderVoidTotemOnChest implements ICurioRenderer {
         ICurioRenderer.translateIfSneaking(poseStack, livingEntity);
         ICurioRenderer.rotateIfSneaking(poseStack, livingEntity);
 
+        BTUConfigs btuConfigs = BetterTotemOfUndying.config;
+
         poseStack.scale(0.35F, 0.35F, 0.35F);
-        poseStack.translate(0.0F, 1.1F, -0.4F);
+        poseStack.translate(0.0F + btuConfigs.charm.xOffset, 1.1F + btuConfigs.charm.yOffset, -0.4F + btuConfigs.charm.zOffset);
         poseStack.mulPose(Direction.DOWN.getRotation());
 
         Minecraft.getInstance().getItemRenderer().renderStatic(itemStack, ItemDisplayContext.FIXED, 15728880, OverlayTexture.NO_OVERLAY, poseStack, multiBufferSource, Minecraft.getInstance().level, 0);