+++ Forge/src/main/java/com/cerbon/better_totem_of_undying/forge/event/BTUClientEventsForge.java
@@ -2,7 +2,7 @@ package com.cerbon.better_totem_of_undying.forge.event;
 
 import com.cerbon.better_totem_of_undying.BetterTotemOfUndying;
 import com.cerbon.better_totem_of_undying.config.BTUConfigs;
-import com.cerbon.better_totem_of_undying.forge.client.RenderVoidTotemOnChest;
+import com.cerbon.better_totem_of_undying.forge.client.RenderTotemOnChest;
 import com.cerbon.better_totem_of_undying.util.BTUConstants;
 import com.cerbon.cerbons_api.api.static_utilities.MiscUtils;
 import me.shedaniel.autoconfig.AutoConfig;
@@ -21,7 +21,7 @@ public class BTUClientEventsForge {
     @SubscribeEvent
     public static void onClientSetup(FMLClientSetupEvent event) {
         if (MiscUtils.isModLoaded(BTUConstants.CURIOS_MOD_ID) && BetterTotemOfUndying.config.charm.displayTotemOnChest)
-            CuriosRendererRegistry.register(Items.TOTEM_OF_UNDYING, RenderVoidTotemOnChest::new);
+            CuriosRendererRegistry.register(Items.TOTEM_OF_UNDYING, RenderTotemOnChest::new);
 
         ModLoadingContext.get().registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class, () -> new ConfigScreenHandler.ConfigScreenFactory((client, parent) -> AutoConfig.getConfigScreen(BTUConfigs.class, parent).get()));
     }